# dashboard/openapi.json
{
  "components": {
    "schemas": {
      "ActivityEventOut": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "at",
          "job_id",
          "kind",
          "detail"
        ],
        "title": "ActivityEventOut",
        "type": "object"
      },
      "ActivityOut": {
        "properties": {
          "default_from": {
            "title": "Default From",
            "type": "string"
          },
          "default_to": {
            "title": "Default To",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActivityEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "max_backfill_days": {
            "title": "Max Backfill Days",
            "type": "integer"
          },
          "queued": {
            "title": "Queued",
            "type": "integer"
          },
          "running": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheduler_available": {
            "title": "Scheduler Available",
            "type": "boolean"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "running",
          "queued",
          "events",
          "scheduler_available",
          "timezone",
          "default_from",
          "default_to",
          "max_backfill_days"
        ],
        "title": "ActivityOut",
        "type": "object"
      },
...
      "AlertRules": {
        "description": "What to alert on. Seeded from `[alerts.rules]` and editable from the\ndashboard, unless a `REOVAULT_ALERTS__RULES__*` env var pins it (same\nscheme as retention).",
        "properties": {
          "coverage": {
            "default": true,
            "title": "Coverage",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "gateway_down": {
            "default": true,
            "title": "Gateway Down",
            "type": "boolean"
          },
          "gateway_down_minutes": {
            "default": 15,
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Gateway Down Minutes",
            "type": "integer"
          },
          "notify_on_resolve": {
            "default": true,
            "title": "Notify On Resolve",
            "type": "boolean"
          },
          "problems": {
            "default": true,
            "title": "Problems",
            "type": "boolean"
          },
          "renotify_hours": {
            "default": 24,
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Renotify Hours",
            "type": "integer"
          },
          "run_failed": {
            "default": true,
            "title": "Run Failed",
            "type": "boolean"
          }
        },
        "title": "AlertRules",
        "type": "object"
      },
      "AlertsOut": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelOut"
            },
            "title": "Channels",
            "type": "array"
          },
          "env_pinned": {
            "title": "Env Pinned",
            "type": "boolean"
          },
          "open": {
            "items": {
              "$ref": "#/components/schemas/OpenAlertOut"
            },
            "title": "Open",
            "type": "array"
          },
          "rules": {
            "$ref": "#/components/schemas/AlertRules"
          }
        },
        "required": [
          "rules",
          "env_pinned",
          "channels",
          "open"
        ],
        "title": "AlertsOut",
        "type": "object"
      },
      "BackfillIn": {
        "properties": {
          "from_local": {
            "title": "From Local",
            "type": "string"
          },
          "to_local": {
            "title": "To Local",
            "type": "string"
          }
        },
        "required": [
          "from_local",
          "to_local"
        ],
        "title": "BackfillIn",
        "type": "object"
      },
...
      "ChannelOut": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "configured"
        ],
        "title": "ChannelOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenAlertOut": {
        "properties": {
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "last_sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sent At"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "condition",
          "device_id",
          "title",
          "message",
          "first_seen_at",
          "last_sent_at"
        ],
        "title": "OpenAlertOut",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "bytes_archived": {
            "title": "Bytes Archived",
            "type": "integer"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "discovered": {
            "title": "Discovered",
            "type": "integer"
          },
          "downloaded": {
            "title": "Downloaded",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "skipped_dup": {
            "title": "Skipped Dup",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "window_from_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window From Utc"
          },
          "window_to_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window To Utc"
          }
        },
        "required": [
          "id",
          "device_id",
          "trigger",
          "window_from_utc",
          "window_to_utc",
          "started_at",
          "finished_at",
          "outcome",
          "discovered",
          "downloaded",
          "skipped_dup",
          "failed",
          "bytes_archived",
          "error"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
      "TestResultOut": {
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "channel",
          "ok",
          "error"
        ],
        "title": "TestResultOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ReoVault",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/actions/backfill": {
      "post": {
        "operationId": "action_backfill_api_v1_actions_backfill_post",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Backfill",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/reconcile": {
      "post": {
        "operationId": "action_reconcile_api_v1_actions_reconcile_post",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Reconcile",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/run": {
      "post": {
        "operationId": "action_run_api_v1_actions_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Run",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/stop": {
      "post": {
        "description": "Cooperative: the Archiver checks the flag between recordings, so a\nfetch or encrypt in flight finishes or fails on its own rather than\nbeing torn down mid-write. The run then closes out as `aborted`.",
        "operationId": "action_stop_api_v1_actions_stop_post",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Stop",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/api/v1/alerts/rules": {
      "put": {
        "operationId": "put_rules_api_v1_alerts_rules_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRules"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Rules",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/v1/alerts/test": {
      "post": {
        "operationId": "send_test_api_v1_alerts_test_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestResultOut"
                  },
                  "title": "Response Send Test Api V1 Alerts Test Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Send Test",
        "tags": [
          "alerts"
        ]
      }
    },
...
  }
}